+++ lib/config.py
@@ -1,10 +1,11 @@
 import json
+import os
 import sys
 from pathlib import Path
 from typing import Any, Dict, List
 
 from .log import get_logger
-from .utils import NotificationLevel
+from .utils import NotificationLevel, is_truthy
 
 # Type alias for JSON
 JSON = Dict[str, Any]
@@ -108,6 +109,7 @@ class GlobalConfig(Config):
 
         try:
             config = self._read_config()
+            config = self._read_env_vars(config)
             self._parse_config(config)
         except (ConfigError, json.decoder.JSONDecodeError) as err:
             print("Error in configuration file:")
@@ -141,6 +143,75 @@ class GlobalConfig(Config):
 
         if not isinstance(config, dict):
             raise ConfigError("Configuration must be a JSON dictionary")
+
+        return config
+
+    def _read_env_vars(self, config: JSON) -> JSON:
+        logger.debug("Reading configuration from environment variables")
+        # Check Fares
+        check_fares = os.getenv("AUTO_SOUTHWEST_CHECK_IN_CHECK_FARES")
+        if check_fares:
+            try:
+                config["check_fares"] = is_truthy(check_fares)
+            except ValueError as err:
+                raise ConfigError("Error parsing 'AUTO_SOUTHWEST_CHECK_IN_CHECK_FARES'") from err
+
+        # Notification URL
+        notification_url = os.getenv("AUTO_SOUTHWEST_CHECK_IN_NOTIFICATION_URL")
+        if notification_url:
+            config.setdefault("notification_urls", [])
+            if isinstance(config["notification_urls"], str):
+                config["notification_urls"] = [config["notification_urls"]]
+            if not isinstance(config["notification_urls"], list):
+                raise ConfigError("'notification_urls' must be a string or a list")
+            if notification_url not in config["notification_urls"]:
+                config["notification_urls"].append(notification_url)
+
+        # Notification Level
+        notification_level = os.getenv("AUTO_SOUTHWEST_CHECK_IN_NOTIFICATION_LEVEL")
+        if notification_level:
+            try:
+                config["notification_level"] = int(notification_level)
+            except ValueError as err:
+                raise ConfigError(
+                    "'AUTO_SOUTHWEST_CHECK_IN_NOTIFICATION_LEVEL' must be an integer"
+                ) from err
+
+        # Browser Path
+        browser_path = os.getenv("AUTO_SOUTHWEST_CHECK_IN_BROWSER_PATH")
+        if browser_path:
+            config["browser_path"] = browser_path
+
+        # Retrieval Interval
+        retrieval_interval = os.getenv("AUTO_SOUTHWEST_CHECK_IN_RETRIEVAL_INTERVAL")
+        if retrieval_interval:
+            try:
+                config["retrieval_interval"] = int(retrieval_interval)
+            except ValueError as err:
+                raise ConfigError(
+                    "'AUTO_SOUTHWEST_CHECK_IN_RETRIEVAL_INTERVAL' must be an integer"
+                ) from err
+
+        # Account credentials
+        username = os.getenv("AUTO_SOUTHWEST_CHECK_IN_USERNAME")
+        password = os.getenv("AUTO_SOUTHWEST_CHECK_IN_PASSWORD")
+        if username and password:
+            new_credentials = {"username": username, "password": password}
+            config.setdefault("accounts", [])
+            config["accounts"].append(new_credentials)
+
+        # Reservation information
+        confirmation_number = os.getenv("AUTO_SOUTHWEST_CHECK_IN_CONFIRMATION_NUMBER")
+        first_name = os.getenv("AUTO_SOUTHWEST_CHECK_IN_FIRST_NAME")
+        last_name = os.getenv("AUTO_SOUTHWEST_CHECK_IN_LAST_NAME")
+        if confirmation_number and first_name and last_name:
+            new_reservation = {
+                "confirmationNumber": confirmation_number,
+                "firstName": first_name,
+                "lastName": last_name,
+            }
+            config.setdefault("reservations", [])
+            config["reservations"].append(new_reservation)
 
         return config