+++ CHANGELOG.md
@@ -4,6 +4,10 @@ If there is no "Upgrading" header for that version, no post-upgrade actions need
 
 
 ## Upcoming
+### New Features
+- Added environment variable alternatives for many configuration items
+([#210](https://github.com/jdholtz/auto-southwest-check-in/pull/210))
+
 ### Bug Fixes
 - Fix failed logins not reporting the correct error
 ([#189](https://github.com/jdholtz/auto-southwest-check-in/issues/189))